+++ datalayer.js
@@ -1,55 +1,48 @@
-const request = require("request-promise");
+const superagent = require("superagent");
 const fs = require("fs");
-const os = require("os");
 const path = require("path");
 const https = require("https");
 
-const { getConfig } = require("./utils/config-loader");
+const { getConfig, getChiaRoot } = require("./utils/config-loader");
 const CONFIG = getConfig();
 
 const getBaseOptions = () => {
-  const homeDir = os.homedir();
+  const chiaRoot = getChiaRoot();
 
   const certFile = path.resolve(
-    `${homeDir}/.chia/mainnet/config/ssl/data_layer/private_data_layer.crt`
+    `${chiaRoot}/config/ssl/data_layer/private_data_layer.crt`
   );
 
   const keyFile = path.resolve(
-    `${homeDir}/.chia/mainnet/config/ssl/data_layer/private_data_layer.key`
+    `${chiaRoot}/config/ssl/data_layer/private_data_layer.key`
   );
 
   const httpsAgent = new https.Agent({
     rejectUnauthorized: false,
   });
 
-  const baseOptions = {
-    method: "POST",
+  return {
     cert: fs.readFileSync(certFile),
     key: fs.readFileSync(keyFile),
-    timeout: 60000,
     agent: httpsAgent,
+    timeout: 300000,
   };
-
-  return baseOptions;
 };
 
 const getStoreIds = async (orgUid) => {
   try {
     const baseOptions = getBaseOptions();
 
-    const fetchOptions = {
-      ...baseOptions,
-      url: `${CONFIG.DATA_LAYER_HOST}/get_owned_stores`,
-      method: "post",
-      body: JSON.stringify({
-        id: orgUid,
-      }),
-      headers: { "Content-Type": "application/json" },
-    };
+    const response = await superagent
+      .post(`${CONFIG.DATA_LAYER_HOST}/get_owned_stores`)
+      .send({ id: orgUid })
+      .set({ "Content-Type": "application/json" })
+      .agent(baseOptions.agent)
+      .key(baseOptions.key)
+      .cert(baseOptions.cert)
+      .timeout(timeout);
 
-    const response = await request(fetchOptions);
-
-    const data = JSON.parse(response);
+    const data = response.body;
 
     if (data.success) {
       return data.store_ids;