+++ CHANGELOG.md
@@ -1,54 +1,156 @@
-## 1.0.1 (2022-11-11)
+## [1.1.4](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.1.3...1.1.4) (2023-07-14)
+
+
+
+## [1.1.3](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.1.2...1.1.3) (2023-07-10)
+
+
+
+## [1.1.2](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.1.1...1.1.2) (2023-06-29)
+
+
+
+## [1.1.1](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.1.0...1.1.1) (2023-06-13)
+
+
+
+# [1.1.0](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.12...1.1.0) (2023-05-23)
+
+
+
+## [1.0.12](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.11...1.0.12) (2023-05-19)
+
+
+
+## [1.0.11](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.10...1.0.11) (2023-05-18)
+
+
+
+## [1.0.10](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.9...1.0.10) (2023-05-18)
 
 
 ### Bug Fixes
 
-* add orgUid to proxy requests ([5ef63c8](https://github.com/Chia-Network/Climate-Portal/commit/5ef63c85f84a1e6c441aabf87213a2de2583c86d))
-* orguid update is not working ([3b8353d](https://github.com/Chia-Network/Climate-Portal/commit/3b8353de47075a5f44b1ba5193a5322e70292fad))
-* tokenize validation ([fcb8072](https://github.com/Chia-Network/Climate-Portal/commit/fcb80727b4bf7e29be58130d67d4eea0a3e16e27))
-* wallet holder ([14e85b4](https://github.com/Chia-Network/Climate-Portal/commit/14e85b4a476117133bffbd1b2117f7ee1b39392b))
+* move api key middleware up the stack so it applies to all requests ([217db14](https://github.com/Chia-Network/climate-tokenization-engine/commit/217db1401102a245711593a4ec20508322cdcc67))
+* remove trailing comma ([487036a](https://github.com/Chia-Network/climate-tokenization-engine/commit/487036abaa4e5b0b5237ffd0798fa6c7959d4fd4))
+* require api key to start server ([0a64c13](https://github.com/Chia-Network/climate-tokenization-engine/commit/0a64c13c8687856791cb72af3392f530359937a9))
+* update api configuration names ([3976b50](https://github.com/Chia-Network/climate-tokenization-engine/commit/3976b50cb7a47036d594b8b018500066ee202df1))
+* update comment ([b7cb389](https://github.com/Chia-Network/climate-tokenization-engine/commit/b7cb38947c897ecb745be17c5133f84a64ece19e))
+* update tokenize driver name ([180b834](https://github.com/Chia-Network/climate-tokenization-engine/commit/180b83446271446a2bac179fb78db0188a268e61))
+* use CHIA_ROOT environment variable ([2ded22a](https://github.com/Chia-Network/climate-tokenization-engine/commit/2ded22ad98acfe92ceb3b9d148dd2820de25594c))
 
 
 ### Features
 
-* add default fee ([a5eb659](https://github.com/Chia-Network/Climate-Portal/commit/a5eb659ebbb16ac20eea5b58cecfd4ea5f0754cb))
-* add hoi validation ([1ad37ce](https://github.com/Chia-Network/Climate-Portal/commit/1ad37ce0131c16f19f3a1d61a51b1549e81fbb80))
-* add home org does not exist error ([889ad46](https://github.com/Chia-Network/Climate-Portal/commit/889ad46ea20482dee466eff24dff0bd0452b8545))
-* add orguid header ([6c60392](https://github.com/Chia-Network/Climate-Portal/commit/6c6039281d9c86280611c80420a14a8498ecb874))
-* add orgUid to all headers ([ea238f5](https://github.com/Chia-Network/Climate-Portal/commit/ea238f5afb881823eb8b0c14b739676cf8056b17))
-* add pagination by adding the pagination params to the proxy ([7a26759](https://github.com/Chia-Network/Climate-Portal/commit/7a267595092254ca4ef5e210aab06c54df42b3c7))
-* add pagination by adding the pagination params to the proxy ([0d07b60](https://github.com/Chia-Network/Climate-Portal/commit/0d07b60aa1ad605e0cca7fadcbef7949081ebfc2))
-* add projects endpoint proxy ([937597c](https://github.com/Chia-Network/Climate-Portal/commit/937597ce83bfb42cdd20b4ee5f853352b9aeccc6))
-* add proxy requests ([5309cfa](https://github.com/Chia-Network/Climate-Portal/commit/5309cfad54c33ed3ca2a714d61fbdd3e3155b476))
-* add tokenize endpoint ([90df4bf](https://github.com/Chia-Network/Climate-Portal/commit/90df4bfe06cc7ccc8dc9027cc28e687c4a81f753))
-* add tokenize endpoint ([1af0619](https://github.com/Chia-Network/Climate-Portal/commit/1af06192d90a07653ac7a2ad2962e607c5e2f63d))
-* add tokenize endpoint ([c45b275](https://github.com/Chia-Network/Climate-Portal/commit/c45b2751826374d57e3e73bebe768ca9674d8fa4))
-* add tokenize endpoint ([ebc94f8](https://github.com/Chia-Network/Climate-Portal/commit/ebc94f8be5a06d5aafc1e3e5cc302e3d740b9a4a))
-* add tokenize endpoint ([f3547f7](https://github.com/Chia-Network/Climate-Portal/commit/f3547f7814fc69bc32222f150d87692570d5e400))
-* add tokenize endpoint ([44869cd](https://github.com/Chia-Network/Climate-Portal/commit/44869cd7dfec5a7b72bfa9c70dab335bb63f9286))
-* add tokenize endpoint ([88fc089](https://github.com/Chia-Network/Climate-Portal/commit/88fc0899fe20cb3c801763f5bfca8f369e7162d7))
-* add tokenize endpoint ([ba0ee80](https://github.com/Chia-Network/Climate-Portal/commit/ba0ee806b99ca763e7e59adea3045153bd71db21))
-* add tokenize endpoint ([0ebbab8](https://github.com/Chia-Network/Climate-Portal/commit/0ebbab89cdd614d667bd5eefe2feb4c35b92c285))
-* add tokenize endpoint ([250e139](https://github.com/Chia-Network/Climate-Portal/commit/250e139f7e72d8ab0f2f75f67d1e1308c72841c9))
-* add yaml ([1ec6417](https://github.com/Chia-Network/Climate-Portal/commit/1ec641704600e3a838521fe303b825702ca3dd5f))
-* add yaml ([3d06d9e](https://github.com/Chia-Network/Climate-Portal/commit/3d06d9ea0146e0dd36b0832719f296e40ecf36c0))
-* config option to update climate warehouse units ([295b297](https://github.com/Chia-Network/Climate-Portal/commit/295b2978b6fe7477ba42b55aaebd7bd3826c2b42))
-* detokenize ([a2429df](https://github.com/Chia-Network/Climate-Portal/commit/a2429df0c69a49a9611ccfe2a3d8007622576c86))
-* detokenize ([43d36b0](https://github.com/Chia-Network/Climate-Portal/commit/43d36b0e88b2744ef6a275c138b7af5c7755533d))
-* get sore ids from datalayer ([8f741ef](https://github.com/Chia-Network/Climate-Portal/commit/8f741efa712db1a7574f68cbf937fbf840fb2fc3))
-* get store ids from datalayer ([6b837c6](https://github.com/Chia-Network/Climate-Portal/commit/6b837c635889747b481eaa7c55f0938059fdb348))
-* get store ids from datalayer ([b0eee6e](https://github.com/Chia-Network/Climate-Portal/commit/b0eee6e17527c06cbea52c39dc01dcec0658cbc4))
-* get store ids from datalayer ([da96dd2](https://github.com/Chia-Network/Climate-Portal/commit/da96dd23e407bae091ae9ec9d7cb085d8bbf0d2c))
-* get store ids from datalayer ([d0d759b](https://github.com/Chia-Network/Climate-Portal/commit/d0d759ba2b6462e8368aed81143a0b8fd819260c))
-* package into executable ([a36c062](https://github.com/Chia-Network/Climate-Portal/commit/a36c062b90990de315863c39c43ec9d8ac06a402))
-* remove default fee ([32494fe](https://github.com/Chia-Network/Climate-Portal/commit/32494fe1b633b5990d876ee348535be6736c6f9a))
-* remove fee field ([ff0b071](https://github.com/Chia-Network/Climate-Portal/commit/ff0b071f83a27cc2a1f3350b60d89d71a503a159))
-* replace detok file with detok string ([5701455](https://github.com/Chia-Network/Climate-Portal/commit/5701455768b2d61522854b49a5ca7b719681533b))
-* replace detok file with detok string ([76a8627](https://github.com/Chia-Network/Climate-Portal/commit/76a862791affba9790da1c707a9aebe47591c643))
-* return data for ui from detok-file endpoint ([2fd3e89](https://github.com/Chia-Network/Climate-Portal/commit/2fd3e899710449593e3b857d46e2b6595cd1f4b9))
-* return data for ui from detok-file endpoint ([66cf8a9](https://github.com/Chia-Network/Climate-Portal/commit/66cf8a95d8dc0c89d775ef05ec6e2e76b7fa9201))
-* send detok confirmation ([89c9d7d](https://github.com/Chia-Network/Climate-Portal/commit/89c9d7de5c0cebceaf4f488c6f967c2d720ee11e))
-* set parse detok unit information to amount in detok file ([0ed694a](https://github.com/Chia-Network/Climate-Portal/commit/0ed694a0ea0bc580842738640e9691a1318a6e60))
+* add persistent logging ([49ecba3](https://github.com/Chia-Network/climate-tokenization-engine/commit/49ecba37b9554b3bdc152ff24d812c502ca3bc42))
+
+
+
+## [1.0.8](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.7...1.0.8) (2023-04-24)
+
+
+### Bug Fixes
+
+* include api key in requests to cadt ([f755b5a](https://github.com/Chia-Network/climate-tokenization-engine/commit/f755b5a65bc12c14ccc3ba9039c06dca42dc199e))
+
+
+
+## [1.0.7](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.6...1.0.7) (2023-04-19)
+
+
+### Bug Fixes
+
+* add UNITS_FILTER to default config ([ba0f4ce](https://github.com/Chia-Network/climate-tokenization-engine/commit/ba0f4ce571b90c321664f182f79ca578da32582f))
+
+
+
+## [1.0.6](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.5...1.0.6) (2023-04-18)
+
+
+### Features
+
+* support api key ([451efca](https://github.com/Chia-Network/climate-tokenization-engine/commit/451efca62430366c118d0a8f89c0ca3c297592b5))
+
+
+
+## [1.0.5](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.4...1.0.5) (2023-01-13)
+
+
+### Features
+
+* filter untokenized units by portal api config param ([0946480](https://github.com/Chia-Network/climate-tokenization-engine/commit/09464809d023ea1f8562dbdd9468031ab616bfec))
+
+
+
+## [1.0.4](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.2...1.0.4) (2023-01-04)
+
+
+### Features
+
+* add org-uid headers to all responses ([77b7ce7](https://github.com/Chia-Network/climate-tokenization-engine/commit/77b7ce7999e7b6af08088fdeb5746b77d8813f0f))
+
+
+
+## [1.0.2](https://github.com/Chia-Network/climate-tokenization-engine/compare/1.0.1...1.0.2) (2022-12-02)
+
+
+### Features
+
+* add default search param ([d5108f9](https://github.com/Chia-Network/climate-tokenization-engine/commit/d5108f99b9d0ea1de30f1a2fa79659ce7fa28238))
+
+
+
+## [1.0.1](https://github.com/Chia-Network/climate-tokenization-engine/compare/5309cfad54c33ed3ca2a714d61fbdd3e3155b476...1.0.1) (2022-11-11)
+
+
+### Bug Fixes
+
+* add orgUid to proxy requests ([5ef63c8](https://github.com/Chia-Network/climate-tokenization-engine/commit/5ef63c85f84a1e6c441aabf87213a2de2583c86d))
+* orguid update is not working ([3b8353d](https://github.com/Chia-Network/climate-tokenization-engine/commit/3b8353de47075a5f44b1ba5193a5322e70292fad))
+* tokenize validation ([fcb8072](https://github.com/Chia-Network/climate-tokenization-engine/commit/fcb80727b4bf7e29be58130d67d4eea0a3e16e27))
+* wallet holder ([14e85b4](https://github.com/Chia-Network/climate-tokenization-engine/commit/14e85b4a476117133bffbd1b2117f7ee1b39392b))
+
+
+### Features
+
+* add default fee ([a5eb659](https://github.com/Chia-Network/climate-tokenization-engine/commit/a5eb659ebbb16ac20eea5b58cecfd4ea5f0754cb))
+* add hoi validation ([1ad37ce](https://github.com/Chia-Network/climate-tokenization-engine/commit/1ad37ce0131c16f19f3a1d61a51b1549e81fbb80))
+* add home org does not exist error ([889ad46](https://github.com/Chia-Network/climate-tokenization-engine/commit/889ad46ea20482dee466eff24dff0bd0452b8545))
+* add orguid header ([6c60392](https://github.com/Chia-Network/climate-tokenization-engine/commit/6c6039281d9c86280611c80420a14a8498ecb874))
+* add orgUid to all headers ([ea238f5](https://github.com/Chia-Network/climate-tokenization-engine/commit/ea238f5afb881823eb8b0c14b739676cf8056b17))
+* add pagination by adding the pagination params to the proxy ([7a26759](https://github.com/Chia-Network/climate-tokenization-engine/commit/7a267595092254ca4ef5e210aab06c54df42b3c7))
+* add pagination by adding the pagination params to the proxy ([0d07b60](https://github.com/Chia-Network/climate-tokenization-engine/commit/0d07b60aa1ad605e0cca7fadcbef7949081ebfc2))
+* add projects endpoint proxy ([937597c](https://github.com/Chia-Network/climate-tokenization-engine/commit/937597ce83bfb42cdd20b4ee5f853352b9aeccc6))
+* add proxy requests ([5309cfa](https://github.com/Chia-Network/climate-tokenization-engine/commit/5309cfad54c33ed3ca2a714d61fbdd3e3155b476))
+* add tokenize endpoint ([90df4bf](https://github.com/Chia-Network/climate-tokenization-engine/commit/90df4bfe06cc7ccc8dc9027cc28e687c4a81f753))
+* add tokenize endpoint ([1af0619](https://github.com/Chia-Network/climate-tokenization-engine/commit/1af06192d90a07653ac7a2ad2962e607c5e2f63d))
+* add tokenize endpoint ([c45b275](https://github.com/Chia-Network/climate-tokenization-engine/commit/c45b2751826374d57e3e73bebe768ca9674d8fa4))
+* add tokenize endpoint ([ebc94f8](https://github.com/Chia-Network/climate-tokenization-engine/commit/ebc94f8be5a06d5aafc1e3e5cc302e3d740b9a4a))
+* add tokenize endpoint ([f3547f7](https://github.com/Chia-Network/climate-tokenization-engine/commit/f3547f7814fc69bc32222f150d87692570d5e400))
+* add tokenize endpoint ([44869cd](https://github.com/Chia-Network/climate-tokenization-engine/commit/44869cd7dfec5a7b72bfa9c70dab335bb63f9286))
+* add tokenize endpoint ([88fc089](https://github.com/Chia-Network/climate-tokenization-engine/commit/88fc0899fe20cb3c801763f5bfca8f369e7162d7))
+* add tokenize endpoint ([ba0ee80](https://github.com/Chia-Network/climate-tokenization-engine/commit/ba0ee806b99ca763e7e59adea3045153bd71db21))
+* add tokenize endpoint ([0ebbab8](https://github.com/Chia-Network/climate-tokenization-engine/commit/0ebbab89cdd614d667bd5eefe2feb4c35b92c285))
+* add tokenize endpoint ([250e139](https://github.com/Chia-Network/climate-tokenization-engine/commit/250e139f7e72d8ab0f2f75f67d1e1308c72841c9))
+* add yaml ([1ec6417](https://github.com/Chia-Network/climate-tokenization-engine/commit/1ec641704600e3a838521fe303b825702ca3dd5f))
+* add yaml ([3d06d9e](https://github.com/Chia-Network/climate-tokenization-engine/commit/3d06d9ea0146e0dd36b0832719f296e40ecf36c0))
+* config option to update climate warehouse units ([295b297](https://github.com/Chia-Network/climate-tokenization-engine/commit/295b2978b6fe7477ba42b55aaebd7bd3826c2b42))
+* detokenize ([a2429df](https://github.com/Chia-Network/climate-tokenization-engine/commit/a2429df0c69a49a9611ccfe2a3d8007622576c86))
+* detokenize ([43d36b0](https://github.com/Chia-Network/climate-tokenization-engine/commit/43d36b0e88b2744ef6a275c138b7af5c7755533d))
+* get sore ids from datalayer ([8f741ef](https://github.com/Chia-Network/climate-tokenization-engine/commit/8f741efa712db1a7574f68cbf937fbf840fb2fc3))
+* get store ids from datalayer ([6b837c6](https://github.com/Chia-Network/climate-tokenization-engine/commit/6b837c635889747b481eaa7c55f0938059fdb348))
+* get store ids from datalayer ([b0eee6e](https://github.com/Chia-Network/climate-tokenization-engine/commit/b0eee6e17527c06cbea52c39dc01dcec0658cbc4))
+* get store ids from datalayer ([da96dd2](https://github.com/Chia-Network/climate-tokenization-engine/commit/da96dd23e407bae091ae9ec9d7cb085d8bbf0d2c))
+* get store ids from datalayer ([d0d759b](https://github.com/Chia-Network/climate-tokenization-engine/commit/d0d759ba2b6462e8368aed81143a0b8fd819260c))
+* package into executable ([a36c062](https://github.com/Chia-Network/climate-tokenization-engine/commit/a36c062b90990de315863c39c43ec9d8ac06a402))
+* remove default fee ([32494fe](https://github.com/Chia-Network/climate-tokenization-engine/commit/32494fe1b633b5990d876ee348535be6736c6f9a))
+* remove fee field ([ff0b071](https://github.com/Chia-Network/climate-tokenization-engine/commit/ff0b071f83a27cc2a1f3350b60d89d71a503a159))
+* replace detok file with detok string ([5701455](https://github.com/Chia-Network/climate-tokenization-engine/commit/5701455768b2d61522854b49a5ca7b719681533b))
+* replace detok file with detok string ([76a8627](https://github.com/Chia-Network/climate-tokenization-engine/commit/76a862791affba9790da1c707a9aebe47591c643))
+* return data for ui from detok-file endpoint ([2fd3e89](https://github.com/Chia-Network/climate-tokenization-engine/commit/2fd3e899710449593e3b857d46e2b6595cd1f4b9))
+* return data for ui from detok-file endpoint ([66cf8a9](https://github.com/Chia-Network/climate-tokenization-engine/commit/66cf8a95d8dc0c89d775ef05ec6e2e76b7fa9201))
+* send detok confirmation ([89c9d7d](https://github.com/Chia-Network/climate-tokenization-engine/commit/89c9d7de5c0cebceaf4f488c6f967c2d720ee11e))
+* set parse detok unit information to amount in detok file ([0ed694a](https://github.com/Chia-Network/climate-tokenization-engine/commit/0ed694a0ea0bc580842738640e9691a1318a6e60))